+++ frame.h
@@ -41,39 +41,6 @@ int gotoxy(int x, int y)
 	return 0;
 }
 
-int welcome (void)
-{
-    char welcome[15][95]={"                  #############                                             ####            ",
-                        "             ###################                                          ####              ",
-                        "          ######################                                        ####                ",
-                        "         ########         ####                                        #####                 ",
-                        "        ####            ##                                           ####                   ",
-                        "        ###                                             ####       ####                 ### ",
-                        "        ######                           ####        ########     ####   #####       #######",
-                        "          ##########             ###   ######     ####    ###   ####   #######     ###   ###",
-                        "              ##########       ####  #######    ####     ###   ###   ###   ##    ###   ###  ",
-                        "                      ####    ###  ###  ##    ###      ###    ###  ###   ##     ###  ##     ",
-                        "                        ###  ### ###   ##    ###     ####    ### ####  ##      #####     #  ",
-                        "  ###               ######    ####   ######  ######## ###### ####   ########   #####   ##   ",
-                        " #######################            ####      ###    ####     #      ####       #######     ",
-                        "   ##################                                                  By weilinfox         ",
-                        "                                          Data loaded successfully...                       "};
-
-    int i, j;
-    int len=strlen(welcome[0]);
-
-    for (i=0; welcome[0][i]!='\0'; i++) {
-        for (j=0; j<15; j++) {
-            gotoxy((MAP_X-len+15)/2+i+1, (MAP_Y-15)/2+j+1);
-            putchar(welcome[j][i]);
-        }
-        Sleep(10);
-    }
-    Sleep(1500);
-
-    return 0;
-}
-
 /** get a frame*/
 int getFrame (char frame[][MAP_X], char map[][MAP_X], snakeBodyNode* node)
 {
